+++ src/Map/src/Polygon.php
@@ -28,14 +28,10 @@ final class Polygon implements Element
      */
     public function __construct(
         private readonly array $points,
-        private readonly ?string $title = null,
         private readonly ?InfoWindow $infoWindow = null,
         private readonly array $extra = [],
         public readonly ?string $id = null,
     ) {
-        if (null !== $title) {
-            trigger_deprecation('symfony/ux-map', '2.30', 'The "title" parameter is deprecated and will be removed in 3.0. Use "infoWindow" instead.');
-        }
     }
 
     /**
@@ -55,7 +51,6 @@ final class Polygon implements Element
             'points' => current($this->points) instanceof Point
                 ? array_map(fn (Point $point) => $point->toArray(), $this->points)
                 : array_map(fn (array $path) => array_map(fn (Point $point) => $point->toArray(), $path), $this->points),
-            'title' => $this->title,
             'infoWindow' => $this->infoWindow?->toArray(),
             'extra' => $this->extra,
             'id' => $this->id,